Remove unnecessary cast
The variables already have the same type. If we cast one operand to int
it would be immediatly recast to unsigned, it's original type.

+++ SourceFile.cpp
@@ -150,7 +150,7 @@ bool SourceFile::isSourceLine(const std::string& line) const {
         }
     }
 
-    bool bRet = ((int)tmp.size() >= m_minChars);
+    bool bRet = (tmp.size() >= m_minChars);
     assert(bRet);
     
     return bRet;
